feat: initialize sqlite db with sequelize
initialize a database with sqlite3 and use sequelize.
initialize sequelize, create models to be able to create the tables.
finally, create a seeders to request all rick and morty episodes and add
them to our database.
create a Makefile to be able to create and initialize the database for
out API (create sqlite db, create tables and seed db with data).